backend/pyproject.toml
[project]
name = "gjo-se-backend"
version = "0.1.0"
description = "gjo-se.com – FastAPI Backend"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.1",
    "greenlet>=3.3.2",
    "bcrypt>=4.0.0",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.48",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "httpx>=0.28.1",
    "pre-commit>=4.5.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.4",
]

# ---------------------------------------------------------------------------
# Ruff – Linting
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

# ---------------------------------------------------------------------------
# Black – Formatierung
# ---------------------------------------------------------------------------
[tool.black]
line-length = 88
target-version = ["py312"]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

# ---------------------------------------------------------------------------
# Pyright – Type Checking
# ---------------------------------------------------------------------------
[tool.pyright]
pythonVersion = "3.12"
strict = true
